Backward compatible change that introduces new structure spawn events to register for.
This makes it more extendable and allows for a future API breaking change
that makes it even more flexible.

There is a new GetStructureTEmplateBlocks that components can handle to
modify the blocks that get placed by a structure template.

This patch adds other events systems that want to contribute to the
structure spawn process can register for.

The StructureSpwnStartedEvent gets send when the structure spawning starts
and StructureBlocksSpawnedEvent gets send when the blocks get placed.

The actual SpawnStructureEvent can then later be made consumable.
In future a system might consume the SpawnStructureEvent and do a
delayed block placement until it sends a StructureBlocksSpawnedEvent.